canvaskit-types - rest of the factories
